+++ tests/Phan/LanguageServer/LanguageServerIntegrationTest.php
@@ -1219,7 +1219,7 @@ EOT
             $cur_line = \explode("\n", $new_file_contents)[$position->line] ?? '';
 
             $message = "Unexpected definition for {$position->line}:{$position->character} (0-based) on line \"" . $cur_line . '"' . ' at "' . \substr($cur_line, $position->character, 10) . '"';
-            $this->assertSame($expected_definition_response, $definition_response, $message);  // slightly better diff view than assertSame
+            $this->assertEquals($expected_definition_response, $definition_response, $message);  // slightly better diff view than assertSame
             $this->assertSame($expected_definition_response, $definition_response, $message);
 
             // This operation should be idempotent.
@@ -1228,7 +1228,7 @@ EOT
             $expected_definition_response['id'] = 3;
 
             $definition_response = $perform_definition_request();
-            $this->assertSame($expected_definition_response, $definition_response, $message);  // slightly better diff view than assertSame
+            $this->assertEquals($expected_definition_response, $definition_response, $message);  // slightly better diff view than assertSame
             $this->assertSame($expected_definition_response, $definition_response, $message);
 
             $this->writeShutdownRequestAndAwaitResponse($proc_in, $proc_out);
@@ -1302,7 +1302,7 @@ EOT
                 (string)\json_encode($cur_line),
                 (string)\substr($cur_line, $position->character, 10)
             );
-            $this->assertSame($expected_definition_response, $definition_response, $message);  // slightly better diff view than assertSame
+            $this->assertEquals($expected_definition_response, $definition_response, $message);  // slightly better diff view than assertSame
             $this->assertSame($expected_definition_response, $definition_response, $message);
 
             // This operation should be idempotent.
@@ -1311,7 +1311,7 @@ EOT
             $expected_definition_response['id'] = 3;
 
             $definition_response = $perform_definition_request();
-            $this->assertSame($expected_definition_response, $definition_response, $message);  // slightly better diff view than assertSame
+            $this->assertEquals($expected_definition_response, $definition_response, $message);  // slightly better diff view than assertSame
             $this->assertSame($expected_definition_response, $definition_response, $message);
 
             $this->writeShutdownRequestAndAwaitResponse($proc_in, $proc_out);
@@ -1391,7 +1391,7 @@ EOT
             $expected_hover_response['id'] = 3;
 
             $hover_response = $perform_hover_request(true);
-            $this->assertSame($expected_hover_response, $hover_response, $message);  // slightly better diff view than assertSame
+            $this->assertEquals($expected_hover_response, $hover_response, $message);  // slightly better diff view than assertSame
             $this->assertSame($expected_hover_response, $hover_response, $message);
 
             $this->writeShutdownRequestAndAwaitResponse($proc_in, $proc_out);
@@ -1694,8 +1694,8 @@ EOT;
             'source' => 'Phan',
             'message' => $expected_message,
         ];
-        // assertSame has a better diff view than assertSame, so run it first.
-        $this->assertSame($expected_diagnostic, $diagnostic);
+        // assertEquals has a better diff view than assertSame, so run it first.
+        $this->assertEquals($expected_diagnostic, $diagnostic);
         $this->assertSame($expected_diagnostic, $diagnostic);
     }