+++ tests/plugin_test/src/139_suppress_empty_statement_warning.php
@@ -31,7 +31,7 @@ foreach ([2] as $_) {
     // TODO implement
 }
 
-foreach ([2] as $i) {
+foreach ([2] as $_) {
     // should warn
 }